Still working on purchases
When you create a purchase or a purchase detail it update the purchase total price but also add the products to the current stock. It means if you had 5 on stock and buy 2 then you have 7 and it will be shown on the product details.
I'm still working on that stuff