Working on user management
Users can be enable or disable and depending of that they can login or not, be selected for new purchases.
User list added, detail view also added to see all the user information.
Splitted models and viewmodels on diferent folders.
I made some little changes on the registration for users. It had an error and also fixed it.